// src/modules/form/constants/errors.enum.ts
export enum ErrorMessage {
  NO_CONTENT = 'Không có dữ liệu hiển thị.',
  ID_NAN_ERROR = 'Giá trị [id] không hợp lệ.',

  FORM_ID_EMPTY_ERROR = 'Bạn vui lòng chọn biểu mẫu.',
  FORM_NOT_FOUND_ERROR = '[Biểu mẫu] không tồn tại (id: %s)',

  HEADER_ID_EMPTY_ERROR = 'Bạn vui lòng chọn hạng mục đánh giá.',
  HEADER_NOT_FOUND_ERROR = '[Hạng mục đánh giá] không tồn tại (id: %s)',

  TITLE_ID_EMPTY_ERROR = 'Bạn vui lòng chọn tiêu chí đánh giá.',

  TIME_NAN_ERROR = 'Thời gian không hợp lệ.',

  ACADEMIC_YEAR_NOT_FOUND_ERROR = '[Niên khóa] không tồn tại (id: %s)',
  SEMESTER_NOT_FOUND_ERROR = '[Học kì] không tồn tại (id: %s)',

  TITLE_NOT_FOUND_ERROR = '[Tiêu chí đánh giá] không tồn tại (id: %s)',

  OPERATOR_FORM_ERROR = 'Lưu thông tin biểu mẫu thất bại.',
  OPERATOR_HEADERS_ERROR = 'Lưu nội dung tiêu đề thất bại.',
  OPERATOR_TITLE_ERROR = 'Lưu nội dung hạng mục đánh giá thất bại.',
  OPERATOR_ITEM_ERROR = 'Lưu nội dung chấm điểm thất bại.',
}